+++ internal/provider/eventgatewayproducepolicyschemavalidation_resource_sdk.go
@@ -271,10 +271,10 @@ func (r *EventGatewayProducePolicySchemaValidationResourceModel) ToSharedEventGa
 		}
 	}
 	labels := make(map[string]*string)
-	for labelsKey, labelsValue := range r.Labels {
+	for labelsKey := range r.Labels {
 		labelsInst := new(string)
-		if !labelsValue.IsUnknown() && !labelsValue.IsNull() {
-			*labelsInst = labelsValue.ValueString()
+		if !r.Labels[labelsKey].IsUnknown() && !r.Labels[labelsKey].IsNull() {
+			*labelsInst = r.Labels[labelsKey].ValueString()
 		} else {
 			labelsInst = nil
 		}