+++ internal/provider/eventgatewayschemaregistry_resource.go
@@ -20,7 +20,6 @@ import (
 	speakeasy_stringplanmodifier "github.com/kong/terraform-provider-konnect-beta/internal/planmodifiers/stringplanmodifier"
 	tfTypes "github.com/kong/terraform-provider-konnect-beta/internal/provider/types"
 	"github.com/kong/terraform-provider-konnect-beta/internal/sdk"
-	"github.com/kong/terraform-provider-konnect-beta/internal/validators"
 )
 
 // Ensure provider defined types fully satisfy framework interfaces.
@@ -151,16 +150,10 @@ func (r *EventGatewaySchemaRegistryResource) Schema(ctx context.Context, req res
 					speakeasy_stringplanmodifier.SuppressDiff(speakeasy_stringplanmodifier.ExplicitSuppress),
 				},
 				Description: `An ISO-8601 timestamp representation of entity creation date.`,
-				Validators: []validator.String{
-					validators.IsRFC3339(),
-				},
 			},
 			"description": schema.StringAttribute{
 				Computed:    true,
 				Description: `A human-readable description of the virtual cluster.`,
-				Validators: []validator.String{
-					stringvalidator.UTF8LengthAtMost(512),
-				},
 			},
 			"gateway_id": schema.StringAttribute{
 				Required:    true,
@@ -180,9 +173,6 @@ func (r *EventGatewaySchemaRegistryResource) Schema(ctx context.Context, req res
 			"name": schema.StringAttribute{
 				Computed:    true,
 				Description: `The unique name of the schema registry.`,
-				Validators: []validator.String{
-					stringvalidator.UTF8LengthBetween(1, 255),
-				},
 			},
 			"type": schema.StringAttribute{
 				Computed:    true,
@@ -194,9 +184,6 @@ func (r *EventGatewaySchemaRegistryResource) Schema(ctx context.Context, req res
 					speakeasy_stringplanmodifier.SuppressDiff(speakeasy_stringplanmodifier.ExplicitSuppress),
 				},
 				Description: `An ISO-8601 timestamp representation of entity update date.`,
-				Validators: []validator.String{
-					validators.IsRFC3339(),
-				},
 			},
 		},
 	}
@@ -432,7 +419,10 @@ func (r *EventGatewaySchemaRegistryResource) Delete(ctx context.Context, req res
 		resp.Diagnostics.AddError("unexpected response from API", fmt.Sprintf("%v", res))
 		return
 	}
-	if res.StatusCode != 204 {
+	switch res.StatusCode {
+	case 204, 404:
+		break
+	default:
 		resp.Diagnostics.AddError(fmt.Sprintf("unexpected response from API. Got an unexpected response code %v", res.StatusCode), debugResponse(res.RawResponse))
 		return
 	}